requirements.txt
flask==2.3.2
requests==2.31.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
geopy==2.4.0
basemap==1.3.8
dnspython==2.4.2
wordcloud==1.9.2
matplotlib-venn==0.11.9
gunicorn==22.0.0
pymongo==4.6.0
python-dotenv==1.0.0